+++ src/euroeval/benchmark_modules/vllm.py
@@ -31,7 +31,7 @@ from ..constants import (
     MAX_LOGPROBS,
     MERGE_TAGS,
     REASONING_MAX_TOKENS,
-    REASONING_TOKEN_SHORTLIST,
+    REASONING_TOKENS,
     TASKS_USING_JSON,
     VLLM_BF16_MIN_CUDA_COMPUTE_CAPABILITY,
 )
@@ -1020,11 +1020,7 @@ def clear_vllm() -> None:
 def get_end_of_reasoning_token_id(
     model: "LLM", tokenizer: "PreTrainedTokenizer", model_id: str
 ) -> int | None:
-    """Get the end of reasoning token ID for a generative model.
-
-    This assumes that the reasoning token is of the form <X> and that the end of
-    reasoning token is </X> (for X being any string without spaces). We disallow the
-    reasoning token to be the same as the beginning-of-sentence token.
+    """Get the end-of-reasoning token ID for a generative model.
 
     Args:
         model:
@@ -1037,6 +1033,58 @@ def get_end_of_reasoning_token_id(
     Returns:
         The end of reasoning token ID, or None if it could not be found.
     """
+    reasoning_tokens_in_vocab = [
+        (bor_token, eor_token)
+        for bor_token, eor_token in REASONING_TOKENS
+        if bor_token in tokenizer.get_vocab() and eor_token in tokenizer.get_vocab()
+    ]
+
+    if not reasoning_tokens_in_vocab:
+        reasoning_tokens_str = ", ".join(
+            [
+                f"{bor_token} and {eor_token}"
+                for bor_token, eor_token in REASONING_TOKENS
+            ]
+        )
+        log_once(
+            f"Model {model_id!r} does not have any of the reasoning tokens "
+            f"{reasoning_tokens_str} in its vocabulary, so assuming it is not a "
+            "reasoning model.",
+            level=logging.INFO,
+        )
+        return None
+
+    # Require that the reasoning token is a special token
+    special_tokens = (
+        [
+            decoder_token.content
+            for decoder_token in tokenizer.added_tokens_decoder.values()
+        ]
+        + [encoder_token for encoder_token in tokenizer.added_tokens_encoder.keys()]
+        + tokenizer.all_special_tokens
+    )
+    special_reasoning_tokens = [
+        (bor_token, eor_token)
+        for (bor_token, eor_token) in reasoning_tokens_in_vocab
+        if bor_token in special_tokens and eor_token in special_tokens
+    ]
+
+    if not special_reasoning_tokens:
+        reasoning_tokens_str = ", ".join(
+            [
+                f"{bor_token} and {eor_token}"
+                for bor_token, eor_token in reasoning_tokens_in_vocab
+            ]
+        )
+        log_once(
+            f"Model {model_id!r} does not have any of the reasoning tokens "
+            f"{reasoning_tokens_str} as special tokens, so assuming it is not a "
+            "reasoning model.",
+            level=logging.INFO,
+        )
+        return None
+
+    # Create a prompt to check if the model uses the reasoning tokens
     if tokenizer.chat_template is None:
         prompt = "What is your name?"
     else:
@@ -1048,119 +1096,80 @@ def get_end_of_reasoning_token_id(
         assert isinstance(templated_prompt, str)
         prompt = templated_prompt
 
-    # Generate a completion and remove the BOS token from it, to not confuse it with the
-    # potential reasoning token
-    model_output = model.generate(
-        prompts=[prompt],
-        sampling_params=SamplingParams(max_tokens=10, temperature=0.0),
-        use_tqdm=False,
+    # Check that the beginning-of-reasoning token is actually used by the model
+    completion = (
+        model.generate(
+            prompts=[prompt],
+            sampling_params=SamplingParams(max_tokens=10, temperature=0.0),
+            use_tqdm=False,
+        )[0]
+        .outputs[0]
+        .text
     )
-    completion = model_output[0].outputs[0].text
-
-    if tokenizer.bos_token is not None:
-        if isinstance(tokenizer.bos_token, str):
-            prompt = prompt.replace(tokenizer.bos_token, "").strip()
-            completion = completion.replace(tokenizer.bos_token, "").strip()
-        elif isinstance(tokenizer.bos_token, list):
-            for bos_token in tokenizer.bos_token:
-                prompt = prompt.replace(bos_token, "").strip()
-                completion = completion.replace(bos_token, "").strip()
-
-    # If it doesn't contain a reasoning token, we can't find the end of reasoning token
-    prompt_matches = re.findall(pattern=r"<\w+>", string=prompt)
-    completion_matches = re.findall(pattern=r"<\w+>", string=completion)
-    if not completion_matches and not prompt_matches:
+    bor_reasoning_matches = [
+        (bor_token, eor_token)
+        for bor_token, eor_token in special_reasoning_tokens
+        if bor_token in prompt or bor_token in completion
+    ]
+    if not bor_reasoning_matches:
         log_once(
-            f"Could not find a reasoning token for model {model_id!r}, so assuming "
+            f"The model {model_id!r} did not generate any beginning-of-reasoning "
+            "tokens in the prompt or the completion, out of the potential "
+            "beginning-of-reasoning tokens "
+            f"{[bor_token for bor_token, _ in special_reasoning_tokens]}. Assuming "
             "the model is not a reasoning model.",
-            level=logging.DEBUG,
+            level=logging.INFO,
         )
         return None
 
-    # Check that the found reasoning token and its associated end-of-reasoning tokens
-    # are both special tokens
-    elif completion_matches:
-        reasoning_token_candidates = completion_matches
-    else:
-        assert prompt_matches
-        reasoning_token_candidates = prompt_matches
-
-    # Get the tokenizer's special tokens
-    special_tokens = [
-        decoder_token.content
-        for decoder_token in tokenizer.added_tokens_decoder.values()
-    ]
-    special_tokens.extend(
-        [encoder_token for encoder_token in tokenizer.added_tokens_encoder.keys()]
+    # Check that the beginning-of-reasoning token is actually used by the model
+    completion = (
+        model.generate(
+            prompts=[prompt],
+            sampling_params=SamplingParams(
+                max_tokens=REASONING_MAX_TOKENS, temperature=0.0
+            ),
+            use_tqdm=False,
+        )[0]
+        .outputs[0]
+        .text
     )
-    special_tokens.extend(tokenizer.all_special_tokens)
-
-    # Remove any reasoning tokens that are not special tokens or do not have an
-    # associated end-of-reasoning token which is also a special token
-    reasoning_token_candidates = [
-        reasoning_token
-        for reasoning_token in reasoning_token_candidates
-        if reasoning_token not in special_tokens
-        or f"</{reasoning_token[1:-1]}>" not in special_tokens
+    eor_reasoning_matches = [
+        (bor_token, eor_token)
+        for bor_token, eor_token in bor_reasoning_matches
+        if eor_token in completion
     ]
-
-    if not reasoning_token_candidates:
+    if not eor_reasoning_matches:
         log_once(
-            f"Could not find a reasoning token for model {model_id!r}, so assuming "
-            "the model is not a reasoning model.",
-            level=logging.DEBUG,
+            f"The model {model_id!r} did not generate any end-of-reasoning tokens in "
+            "the prompt or the completion, out of the potential end-of-reasoning "
+            f"tokens {[eor_token for _, eor_token in bor_reasoning_matches]}. "
+            "Assuming the model is not a reasoning model.",
+            level=logging.INFO,
         )
         return None
 
-    # If there are multiple reasoning tokens, we prioritise the ones that are in the
-    # shortlist of reasoning tokens, as these are the most likely to be the correct ones
-    if len(reasoning_token_candidates) > 1:
-        prioritised_reasoning_tokens = [
-            reasoning_token
-            for reasoning_token in reasoning_token_candidates
-            if re.search(
-                pattern="|".join(REASONING_TOKEN_SHORTLIST), string=reasoning_token
-            )
-            is not None
-        ]
-        if len(prioritised_reasoning_tokens) > 1:
-            log_once(
-                f"Found multiple reasoning tokens {prioritised_reasoning_tokens} for "
-                f"model {model_id!r}. Using {prioritised_reasoning_tokens[0]!r} as "
-                "the reasoning token. If this is not the correct reasoning token, "
-                "please report this issue.",
-                level=logging.INFO,
-            )
-            reasoning_token_candidates = prioritised_reasoning_tokens[:1]
-        elif not prioritised_reasoning_tokens:
-            log_once(
-                f"Found multiple reasoning tokens {reasoning_token_candidates!r} for "
-                f"model {model_id!r}, but none of them are in the shortlist of "
-                "reasoning tokens. Using the first one as the reasoning token. If "
-                "this is not the correct reasoning token, please report this issue.",
-                level=logging.INFO,
-            )
-            reasoning_token_candidates = reasoning_token_candidates[:1]
-
-    if len(reasoning_token_candidates) != 1:
-        raise InvalidModel(
-            f"Found {len(reasoning_token_candidates)} reasoning tokens "
-            f"{reasoning_token_candidates!r} for model {model_id!r}. Please report "
-            "this issue."
+    if len(eor_reasoning_matches) > 1:
+        log_once(
+            f"Found multiple reasoning tokens {eor_reasoning_matches} for "
+            f"model {model_id!r}. Using {eor_reasoning_matches[0]!r} as "
+            "the reasoning token. If this is not the correct reasoning token, "
+            "please report this issue.",
+            level=logging.INFO,
         )
 
-    reasoning_token = reasoning_token_candidates[0]
-    end_of_reasoning_token = f"</{reasoning_token[1:-1]}>"
-
+    bor_token, eor_token = eor_reasoning_matches[0]
     log_once(
-        f"Detected reasoning token {reasoning_token!r} and end-of-reasoning "
-        f"token {end_of_reasoning_token!r} for model {model_id!r}.",
-        level=logging.DEBUG,
+        f"Detected beginning-of-reasoning token {bor_token!r} and end-of-reasoning "
+        f"token {eor_token!r} for model {model_id!r}.",
+        level=logging.INFO,
     )
 
     # Encode the end of reasoning token and return its ID
-    end_of_reasoning_token_id = tokenizer.encode(
-        text=end_of_reasoning_token, add_special_tokens=False
-    )[0]
+    eor_token_id = tokenizer.convert_tokens_to_ids(tokens=eor_token)
+    assert isinstance(eor_token_id, int), (
+        f"The end of reasoning token {eor_token!r} for model {model_id!r} "
+        "could not be converted to an ID."
+    )
 
-    return end_of_reasoning_token_id
+    return eor_token_id