+++ src/euroeval/constants.py
@@ -65,5 +65,9 @@ MERGE_TAGS = ["merge", "mergekit"]
 # The minimum required CUDA compute capability for using bfloat16 in vLLM
 VLLM_BF16_MIN_CUDA_COMPUTE_CAPABILITY = 8.0
 
-# Used to identify reasoning tokens for tokenizers with multiple candidates
-REASONING_TOKEN_SHORTLIST = ["think", "reason"]
+# Used to detect whether a model is a reasoning model
+REASONING_TOKENS = [
+    ("<think>", "</think>"),
+    ("<reason>", "</reason>"),
+    ("<reasoning>", "</reasoning>"),
+]