+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Allow a model to not have any BOS and EOS tokens.
 - Improved detection of beginning-of-reasoning tokens for models.
+- Improves detection of reasoning tokens, by having a more strict list of possible
+  such tokens.
 
 
 ## [v15.9.1] - 2025-06-01