+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+
+
+## [v15.9.2] - 2025-06-04
 ### Fixed
 - Allow a model to not have any BOS and EOS tokens.
 - Improved detection of beginning-of-reasoning tokens for models.